Removed short-test target
