+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -24,8 +24,8 @@ unable or unwilling to run these tests in your own account, we can run the tests
 and provide test results.
 -->
 - [ ] Builds (`make release`)
-- [ ] Unit tests (`make short-test`) pass
-- [ ] Integration tests (`make test` and `make run-integ-tests`) pass
+- [ ] Unit tests (`make test`) pass
+- [ ] Integration tests (`make run-integ-tests`) pass
 - [ ] Functional tests (`make run-functional-tests`) pass
 
 New tests cover the changes: <!-- yes|no -->
