+++ agent/functional_tests/util/compare_versions_test.go
@@ -1,4 +1,4 @@
-// +build functional !unit
+// +build functional
 // Copyright 2014-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"). You may
