+++ agent/stats/utils_test.go
@@ -1,3 +1,4 @@
+//+build !integration
 // Copyright 2014-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"). You may
